Design: 채팅방 목록 디자인 임시 반영
